More refactoring of helper methods
